Create ChoiceChosenConstraintStrategy
