+++ src/assets/wise5/common/constraint/EvaluateConstraintContext.ts
@@ -1,10 +1,18 @@
+import { ComponentServiceLookupService } from '../../services/componentServiceLookupService';
 import { StudentDataService } from '../../services/studentDataService';
 import { ConstraintStrategy } from './strategies/ConstraintStrategy';
 
 export class EvaluateConstraintContext {
   private strategy: ConstraintStrategy;
 
-  constructor(private dataService: StudentDataService) {}
+  constructor(
+    private componentServiceLookupService: ComponentServiceLookupService,
+    private dataService: StudentDataService
+  ) {}
+
+  getComponentServiceLookupService(): ComponentServiceLookupService {
+    return this.componentServiceLookupService;
+  }
 
   getDataService(): StudentDataService {
     return this.dataService;
